Split event stream code into two files
